+++ scala/ppml/src/main/scala/com/intel/analytics/bigdl/ppml/PPMLContext.scala
@@ -18,7 +18,7 @@ package com.intel.analytics.bigdl.ppml
 
 import com.intel.analytics.bigdl.dllib.NNContext.{checkScalaVersion, checkSparkVersion, createSparkConf, initConf, initNNContext}
 import com.intel.analytics.bigdl.dllib.utils.Log4Error
-import com.intel.analytics.bigdl.ppml.crypto.{AES_CBC_PKCS5PADDING, BigDLEncrypt, Crypto, CryptoMode, DECRYPT, ENCRYPT, EncryptRuntimeException, PLAIN_TEXT, Encrypter}
+import com.intel.analytics.bigdl.ppml.crypto.{AES_CBC_PKCS5PADDING, BigDLEncrypt, Crypto, CryptoMode, DECRYPT, ENCRYPT, EncryptRuntimeException, PLAIN_TEXT}
 import com.intel.analytics.bigdl.ppml.utils.Supportive
 import com.intel.analytics.bigdl.ppml.kms.common.{KeyLoader, KeyLoaderManagement}
 import org.apache.spark.{SparkConf, SparkContext}
@@ -260,7 +260,7 @@ object PPMLContext{
     val conf = sparkSession.sparkContext.getConf
     sparkSession.sparkContext.hadoopConfiguration.set(
       "spark.bigdl.encryter.type",
-      conf.get("spark.bigdl.encryter.type", Encrypter.COMMON))
+      conf.get("spark.bigdl.encryter.type", BigDLEncrypt.COMMON))
     val primaryKeyNames = getPrimaryKeyNames(conf)
     primaryKeyNames.foreach{
       primaryKeyName => {
