+++ scala/ppml/src/main/scala/com/intel/analytics/bigdl/ppml/crypto/CryptoCodec.scala
@@ -110,8 +110,8 @@ object CryptoCodec {
         cryptoMode: CryptoMode,
         conf: Configuration) extends DecompressorStream(in) {
     buffer = new Array[Byte](bufferSize)
-    val encrypterType = conf.get("spark.bigdl.encryter.type", Encrypter.COMMON)
-    val bigdlEncrypt = Encrypter(encrypterType)
+    val encrypterType = conf.get("spark.bigdl.encryter.type", BigDLEncrypt.COMMON)
+    val bigdlEncrypt = BigDLEncrypt(encrypterType)
     var headerVerified = false
 
     override def decompress(b: Array[Byte], off: Int, len: Int): Int = {
@@ -120,22 +120,21 @@ object CryptoCodec {
         eof = true
         return -1
       }
-
-      encrypterType match {
-        case Encrypter.NATIVE_AES_CBC =>
-          // data file is encrypted by native AES cihper
-          // no Mac integrity check defaultly
-          val (_, initializationVector) = bigdlEncrypt.getHeader(in)
-          val ivStr = new String(initializationVector, StandardCharsets.UTF_8)
-          val dataKeyPlainText = conf.get(s"bigdl.read.dataKey.$ivStr.plainText")
-          bigdlEncrypt.init(AES_CBC_PKCS5PADDING, DECRYPT, dataKeyPlainText, initializationVector)
-        case Encrypter.COMMON =>
-          // data file is encrypted by PPML cipher
-          if (!headerVerified) {
+      if (!headerVerified) {
+        headerVerified = true
+        encrypterType match {
+          case BigDLEncrypt.NATIVE_AES_CBC =>
+            // data file is encrypted by native AES cihper
+            // no Mac integrity check defaultly
+            val (_, initializationVector) = bigdlEncrypt.getHeader(in)
+            val ivStr = new String(initializationVector, StandardCharsets.UTF_8)
+            val dataKeyPlainText = conf.get(s"bigdl.read.dataKey.$ivStr.plainText")
+            bigdlEncrypt.init(AES_CBC_PKCS5PADDING, DECRYPT, dataKeyPlainText, initializationVector)
+          case BigDLEncrypt.COMMON =>
+            // data file is encrypted by PPML cipher
             val (encryptedDataKey, initializationVector) = bigdlEncrypt.getHeader(in)
             val dataKeyPlainText = conf.get(s"bigdl.read.dataKey.$encryptedDataKey.plainText")
             bigdlEncrypt.init(cryptoMode, DECRYPT, dataKeyPlainText, initializationVector)
-            headerVerified = true
           }
       }
       val decompressed = bigdlEncrypt.decryptPart(in, buffer)
